update for the CRUD of the Contrat

+++ src/main/java/tn/esprit/spring/entities/Contrat.java
@@ -24,7 +24,7 @@ public class Contrat implements Serializable {
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	//@SequenceGenerator(name="seq")
-	private int reference;
+	private Long reference;
 	
 	@Temporal(TemporalType.DATE)
 	private Date dateDebut;
@@ -61,11 +61,11 @@ public class Contrat implements Serializable {
 		this.dateDebut = dateDebut;
 	}
 
-	public int getReference() {
+	public Long getReference() {
 		return reference;
 	}
 
-	public void setReference(int reference) {
+	public void setReference(Long reference) {
 		this.reference = reference;
 	}
